+++ great_expectations/data_context/store/util.py
@@ -12,69 +12,6 @@ from great_expectations.data_context.util import build_store_from_config
 logger = logging.getLogger(__name__)
 
 
-def build_in_memory_store_backend(
-    module_name: str = "great_expectations.data_context.store",
-    class_name: str = "InMemoryStoreBackend",
-    **kwargs,
-) -> StoreBackend:
-    logger.debug("Starting data_context/store/util.py#build_in_memory_store_backend")
-    store_backend_config: dict = {"module_name": module_name, "class_name": class_name}
-    store_backend_config.update(**kwargs)
-    return build_store_from_config(
-        store_config=store_backend_config,
-        module_name=module_name,
-        runtime_environment=None,
-    )
-
-
-def build_tuple_filesystem_store_backend(
-    base_directory: str,
-    *,
-    module_name: str = "great_expectations.data_context.store",
-    class_name: str = "TupleFilesystemStoreBackend",
-    **kwargs,
-) -> StoreBackend:
-    logger.debug(
-        f"""Starting data_context/store/util.py#build_tuple_filesystem_store_backend using base_directory:
-"{base_directory}"""
-    )
-    store_backend_config: dict = {
-        "module_name": module_name,
-        "class_name": class_name,
-        "base_directory": base_directory,
-    }
-    store_backend_config.update(**kwargs)
-    return build_store_from_config(
-        store_config=store_backend_config,
-        module_name=module_name,
-        runtime_environment=None,
-    )
-
-
-def build_tuple_s3_store_backend(
-    bucket: str,
-    *,
-    module_name: str = "great_expectations.data_context.store",
-    class_name: str = "TupleS3StoreBackend",
-    **kwargs,
-) -> StoreBackend:
-    logger.debug(
-        f"""Starting data_context/store/util.py#build_tuple_s3_store_backend using bucket: {bucket}
-        """
-    )
-    store_backend_config: dict = {
-        "module_name": module_name,
-        "class_name": class_name,
-        "bucket": bucket,
-    }
-    store_backend_config.update(**kwargs)
-    return build_store_from_config(
-        store_config=store_backend_config,
-        module_name=module_name,
-        runtime_environment=None,
-    )
-
-
 def build_configuration_store(
     configuration_class: Any,
     store_name: str,
@@ -188,9 +125,7 @@ def save_checkpoint_config_to_store_backend(
 
 
 def load_checkpoint_config_from_store_backend(
-    store_name: str,
-    store_backend: Union[StoreBackend, dict],
-    checkpoint_name: str,
+    store_name: str, store_backend: Union[StoreBackend, dict], checkpoint_name: str,
 ) -> CheckpointConfig:
     try:
         checkpoint_config: CheckpointConfig = cast(
@@ -211,119 +146,11 @@ def load_checkpoint_config_from_store_backend(
 
 
 def delete_checkpoint_config_from_store_backend(
-    store_name: str,
-    store_backend: Union[StoreBackend, dict],
-    checkpoint_name: str,
+    store_name: str, store_backend: Union[StoreBackend, dict], checkpoint_name: str,
 ):
     delete_config_from_store_backend(
         configuration_class=CheckpointConfig,
         store_name=store_name,
         store_backend=store_backend,
         configuration_key=checkpoint_name,
-    )
-
-
-def save_checkpoint_config_to_filesystem(
-    store_name: str,
-    base_directory: str,
-    checkpoint_name: str,
-    checkpoint_configuration: CheckpointConfig,
-):
-    store_config: dict = {"base_directory": base_directory}
-    store_backend_obj: StoreBackend = build_tuple_filesystem_store_backend(
-        **store_config
-    )
-    save_checkpoint_config_to_store_backend(
-        store_name=store_name,
-        store_backend=store_backend_obj,
-        checkpoint_name=checkpoint_name,
-        checkpoint_configuration=checkpoint_configuration,
-    )
-
-
-def load_checkpoint_config_from_filesystem(
-    store_name: str,
-    base_directory: str,
-    checkpoint_name: str,
-) -> CheckpointConfig:
-    store_config: dict = {"base_directory": base_directory}
-    store_backend_obj: StoreBackend = build_tuple_filesystem_store_backend(
-        **store_config
-    )
-    return load_checkpoint_config_from_store_backend(
-        store_name=store_name,
-        store_backend=store_backend_obj,
-        checkpoint_name=checkpoint_name,
-    )
-
-
-def delete_checkpoint_config_from_filesystem(
-    store_name: str,
-    base_directory: str,
-    checkpoint_name: str,
-):
-    store_config: dict = {"base_directory": base_directory}
-    store_backend_obj: StoreBackend = build_tuple_filesystem_store_backend(
-        **store_config
-    )
-    delete_checkpoint_config_from_store_backend(
-        store_name=store_name,
-        store_backend=store_backend_obj,
-        checkpoint_name=checkpoint_name,
-    )
-
-
-def save_config_to_filesystem(
-    configuration_class: Any,
-    store_name: str,
-    base_directory: str,
-    configuration_key: str,
-    configuration: BaseYamlConfig,
-):
-    store_config: dict = {"base_directory": base_directory}
-    store_backend_obj: StoreBackend = build_tuple_filesystem_store_backend(
-        **store_config
-    )
-    save_config_to_store_backend(
-        configuration_class=configuration_class,
-        store_name=store_name,
-        store_backend=store_backend_obj,
-        configuration_key=configuration_key,
-        configuration=configuration,
-    )
-
-
-def load_config_from_filesystem(
-    configuration_class: Any,
-    store_name: str,
-    base_directory: str,
-    configuration_key: str,
-) -> BaseYamlConfig:
-    store_config: dict = {"base_directory": base_directory}
-    store_backend_obj: StoreBackend = build_tuple_filesystem_store_backend(
-        **store_config
-    )
-    return load_config_from_store_backend(
-        configuration_class=configuration_class,
-        store_name=store_name,
-        store_backend=store_backend_obj,
-        configuration_key=configuration_key,
-    )
-
-
-def delete_config_from_filesystem(
-    configuration_class: Any,
-    store_name: str,
-    base_directory: str,
-    configuration_key: str,
-):
-    store_config: dict = {"base_directory": base_directory}
-    store_backend_obj: StoreBackend = build_tuple_filesystem_store_backend(
-        **store_config
-    )
-    delete_config_from_store_backend(
-        configuration_class=configuration_class,
-        store_name=store_name,
-        store_backend=store_backend_obj,
-        configuration_key=configuration_key,
     )
